worker/requirements.txt
evaluate==0.4.1
# Flask==2.2.2
# Flask==3.0.3
langchain==0.1.16
pika==1.3.2
python-dotenv==1.0.1
spacy==3.7.3
torch==2.3.0
# transformers==4.37.1
transformers==4.41.2